+++ src/rpc/net.cpp
@@ -111,27 +111,29 @@ void NodeStatsToJSON(const CNodeStats& stats, UniValue& obj)
 
     UniValue sendPerMsgCmd(UniValue::VOBJ);
     for (const mapMsgCmdSize::value_type &i : stats.mapSendBytesPerMsgCmd) {
-        if (i.second > 0)
+        if (i.second > 0) {
             sendPerMsgCmd.push_back(Pair(i.first, i.second));
+        }
     }
     obj.push_back(Pair("bytessent_per_msg", sendPerMsgCmd));
 
     UniValue recvPerMsgCmd(UniValue::VOBJ);
     for (const mapMsgCmdSize::value_type &i : stats.mapRecvBytesPerMsgCmd) {
-        if (i.second > 0)
+        if (i.second > 0) {
             recvPerMsgCmd.push_back(Pair(i.first, i.second));
+        }
     }
     obj.push_back(Pair("bytesrecv_per_msg", recvPerMsgCmd));
 }
 
 UniValue getpeerinfo(const JSONRPCRequest& request)
 {
-    if (request.fHelp || request.params.size() >= 2)
+    if (request.fHelp || request.params.size() > 1)
         throw std::runtime_error(
             "getpeerinfo\n"
             "\nReturns data about connected network nodes as a json array of objects.\n"
             "\nArguments:\n"
-            "1. \"node\"         (numeric, optional) Only return information about this specified node id\n"
+            "1. \"id\"         (numeric, optional) Only return information about the peer with the specified id.\n"
             "\nResult:\n"
             "[\n"
             "  {\n"
@@ -189,12 +191,11 @@ UniValue getpeerinfo(const JSONRPCRequest& request)
         CNodeStats stats;
 
         int node_id = request.params[0].get_int();
-        if (g_connman->GetNodeStats(node_id, stats)) {
-            NodeStatsToJSON(stats, obj);
-            ret.push_back(obj);
-        } else {
-            throw JSONRPCError(RPC_INVALID_PARAMETER, strprintf("Node id %u not found", node_id));
+        if (!g_connman->GetNodeStats(node_id, stats)) {
+            throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, strprintf("Node id %u not found", node_id));
         }
+        NodeStatsToJSON(stats, obj);
+        ret.push_back(obj);
     } else {
         std::vector<CNodeStats> vstats;
         g_connman->GetNodeStats(vstats);
@@ -212,47 +213,49 @@ UniValue getpeerinfo(const JSONRPCRequest& request)
 
 UniValue updatepeer(const JSONRPCRequest& request)
 {
-    if (request.fHelp || request.params.size() > 3 || request.params.size() == 0)
+    if (request.fHelp || request.params.size() < 1 || request.params.size() > 3)
         throw std::runtime_error(
             "updatepeer\n"
-            "\nUpdate settings for a network node (debug method).\n"
-            "\nReturns updated information about the node. See getpeerinfo() for the format of the returned object."
+            "\nUpdate settings for a network peer (debug method).\n"
+            "\nReturns updated information about the peer. See getpeerinfo() for the format of the returned object."
             "\nArguments:\n"
-            "1. \"node\"              (string, required) The node id (see getpeerinfo for a list of nodes with their ids)\n"
-            "2. \"whitelisted\"       (bool, optional) whether the node is whitelisted\n"
-            "3. \"manual_connection\" (bool, optional) whether the node was manually added\n"
+            "1. \"id\"                (string, required) The peer id (see getpeerinfo for a list of peers with their ids)\n"
+            "2. \"whitelisted\"       (bool, optional) whether the peer is whitelisted\n"
+            "3. \"manual_connection\" (bool, optional) whether the peer was manually added\n"
             "\nExamples:\n"
             + HelpExampleCli("updatepeer", "0 true")
             + HelpExampleRpc("updatepeer", "0, true")
         );
 
-    if(!g_connman) {
+    if (!g_connman) {
         throw JSONRPCError(RPC_CLIENT_P2P_DISABLED, "Error: Peer-to-peer functionality missing or disabled");
     }
 
     LOCK(cs_main);
 
     RPCTypeCheckArgument(request.params[0], UniValue::VNUM);
-    int nodeid = request.params[0].get_int();
+    int node_id = request.params[0].get_int();
+
+    CNodeStats stats;
+    if (!g_connman->GetNodeStats(node_id, stats)) {
+        throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, strprintf("Node id %u not found", node_id));
+    }
 
     if (!request.params[1].isNull()) {
         RPCTypeCheckArgument(request.params[1], UniValue::VBOOL);
-        if (!g_connman->SetWhitelisted(nodeid, request.params[1].isTrue())) {
-            throw JSONRPCError(RPC_MISC_ERROR, "Failed to update node whitelisting");
+        if (!g_connman->SetWhitelisted(node_id, request.params[1].isTrue())) {
+            throw JSONRPCError(RPC_MISC_ERROR, "Failed to update peer whitelisting");
         }
     }
 
     if (!request.params[2].isNull()) {
         RPCTypeCheckArgument(request.params[2], UniValue::VBOOL);
-        if (!g_connman->SetManualConnection(nodeid, request.params[2].isTrue())) {
-            throw JSONRPCError(RPC_MISC_ERROR, "Failed to update node manual_connection");
+        if (!g_connman->SetManualConnection(node_id, request.params[2].isTrue())) {
+            throw JSONRPCError(RPC_MISC_ERROR, "Failed to update peer manual_connection");
         }
     }
 
-    CNodeStats stats;
-    if (!g_connman->GetNodeStats(nodeid, stats)) {
-        throw JSONRPCError(RPC_INVALID_PARAMETER, strprintf("Node id %u not found", nodeid));
-    }
+    g_connman->GetNodeStats(node_id, stats);
 
     UniValue entry(UniValue::VOBJ);
     NodeStatsToJSON(stats, entry);
@@ -698,7 +701,7 @@ static const CRPCCommand commands[] =
   //  --------------------- ------------------------  -----------------------  ----------
     { "network",            "getconnectioncount",     &getconnectioncount,     {} },
     { "network",            "ping",                   &ping,                   {} },
-    { "network",            "getpeerinfo",            &getpeerinfo,            {"node"} },
+    { "network",            "getpeerinfo",            &getpeerinfo,            {"id"} },
     { "network",            "addnode",                &addnode,                {"node","command"} },
     { "network",            "disconnectnode",         &disconnectnode,         {"address", "nodeid"} },
     { "network",            "getaddednodeinfo",       &getaddednodeinfo,       {"node"} },
@@ -708,7 +711,7 @@ static const CRPCCommand commands[] =
     { "network",            "listbanned",             &listbanned,             {} },
     { "network",            "clearbanned",            &clearbanned,            {} },
     { "network",            "setnetworkactive",       &setnetworkactive,       {"state"} },
-    { "hidden",             "updatepeer",             &updatepeer,             {"node", "whitelisted", "manual_connection"} },
+    { "hidden",             "updatepeer",             &updatepeer,             {"id", "whitelisted", "manual_connection"} },
 };
 
 void RegisterNetRPCCommands(CRPCTable &t)